ui style changes

+++ app/src/main/java/com/naumov/worldweather/domain/weather/WeeklyForecast.kt
@@ -2,7 +2,7 @@ package com.naumov.worldweather.domain.weather
 
 data class WeeklyForecast(
     val date: String,
-    val dayTemperature: Double,
-    val nightTemperature: Double,
+    val dayTemperature: Int,
+    val nightTemperature: Int,
     val weatherType: WeatherType
 )
